refactor: 웹 조작 로직 분리
